+++ src/main/java/uk/gov/ons/census/fwmt/csvservice/service/ccs/CCSGatewayEventsConfig.java
@@ -1,0 +1,17 @@
+package uk.gov.ons.census.fwmt.csvservice.service.ccs;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import uk.gov.ons.census.fwmt.events.component.GatewayEventManager;
+
+@Configuration
+public class CCSGatewayEventsConfig {
+
+  public static final String CSV_CE_REQUEST_EXTRACTED = "CSV_CE_REQUEST_EXTRACTED";
+  public static final String CANONICAL_CCS_CREATE_SENT = "CANONICAL_CCS_CREATE_SENT";
+
+  @Bean
+  public void addCEEvents(GatewayEventManager gatewayEventManager) {
+    gatewayEventManager.addEventTypes(new String[] {CANONICAL_CCS_CREATE_SENT, CSV_CE_REQUEST_EXTRACTED});
+  }
+}
