+++ src/main/java/uk/gov/ons/census/fwmt/csvservice/service/ce/CEConverterServiceImpl.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.census.fwmt.csvservice.service.impl;
+package uk.gov.ons.census.fwmt.csvservice.service.ce;
 
 import com.opencsv.bean.CsvToBean;
 import com.opencsv.bean.CsvToBeanBuilder;
@@ -9,11 +9,9 @@ import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import uk.gov.ons.census.fwmt.canonical.v1.CreateFieldWorkerJobRequest;
 import uk.gov.ons.census.fwmt.common.error.GatewayException;
-import uk.gov.ons.census.fwmt.csvservice.canonical.CanonicalJobHelper;
+import uk.gov.ons.census.fwmt.csvservice.adapter.GatewayActionAdapter;
 import uk.gov.ons.census.fwmt.csvservice.config.GatewayEventsConfig;
-import uk.gov.ons.census.fwmt.csvservice.dto.CCSPropertyListing;
 import uk.gov.ons.census.fwmt.csvservice.dto.CEJobListing;
-import uk.gov.ons.census.fwmt.csvservice.service.CSVAdapterService;
 import uk.gov.ons.census.fwmt.csvservice.service.CSVConverterService;
 import uk.gov.ons.census.fwmt.events.component.GatewayEventManager;
 
@@ -22,34 +20,36 @@ import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 
-import static uk.gov.ons.census.fwmt.csvservice.config.GatewayEventsConfig.CANONICAL_CCS_CREATE_SENT;
-import static uk.gov.ons.census.fwmt.csvservice.config.GatewayEventsConfig.CSV_CCS_REQUEST_EXTRACTED;
+import static uk.gov.ons.census.fwmt.csvservice.config.GatewayEventsConfig.CANONICAL_CE_CREATE_SENT;
+import static uk.gov.ons.census.fwmt.csvservice.config.GatewayEventsConfig.CSV_CE_REQUEST_EXTRACTED;
+import static uk.gov.ons.census.fwmt.csvservice.service.ce.CECanonicalBuilder.createCEJob;
 import static uk.gov.ons.census.fwmt.csvservice.utils.CsvServiceUtils.moveCsvFile;
 
 @Slf4j
 @Service("CE")
-public class CCSConverterServiceImpl implements CSVConverterService {
-  @Value("${gcpBucket.ccslocation}")
+public class CEConverterServiceImpl implements CSVConverterService {
+
+  @Value("${gcpBucket.celocation}")
   private Resource csvGCPFile;
 
-  @Value("${gcpBucket.ccslocation}")
+  @Value("${gcpBucket.celocation}")
   private Path csvPath;
 
-  @Value("${gcpBucket.ccsProcessedPath}")
+  @Value("${gcpBucket.ceProcessedPath}")
   private Path processedPath;
 
   @Autowired
-  private CSVAdapterService csvAdapterService;
+  private GatewayActionAdapter gatewayActionAdapter;
 
   @Autowired
   private GatewayEventManager gatewayEventManager;
 
   @Override
   public void convertCSVToCanonical() throws GatewayException {
-    CsvToBean<CCSPropertyListing> csvToBean;
+    CsvToBean<CEJobListing> csvToBean;
     try {
       csvToBean = new CsvToBeanBuilder(new InputStreamReader(csvGCPFile.getInputStream(), StandardCharsets.UTF_8))
-          .withType(CCSPropertyListing.class)
+          .withType(CEJobListing.class)
           .build();
 
     } catch (IOException e) {
@@ -58,11 +58,11 @@ public class CCSConverterServiceImpl implements CSVConverterService {
       throw new GatewayException(GatewayException.Fault.SYSTEM_ERROR, e, msg);
     }
 
-    for (CCSPropertyListing ccsPropertyListing : csvToBean) {
-      CreateFieldWorkerJobRequest createFieldWorkerJobRequest = CanonicalJobHelper.createCCSJob(ccsPropertyListing);
-      csvAdapterService.sendJobRequest(createFieldWorkerJobRequest, CANONICAL_CCS_CREATE_SENT);
+    for (CEJobListing CEJobListing : csvToBean) {
+      CreateFieldWorkerJobRequest createFieldWorkerJobRequest = createCEJob(CEJobListing);
+      gatewayActionAdapter.sendJobRequest(createFieldWorkerJobRequest, CANONICAL_CE_CREATE_SENT);
       gatewayEventManager
-          .triggerEvent(String.valueOf(createFieldWorkerJobRequest.getCaseId()), CSV_CCS_REQUEST_EXTRACTED);
+          .triggerEvent(String.valueOf(createFieldWorkerJobRequest.getCaseId()), CSV_CE_REQUEST_EXTRACTED);
     }
     moveCsvFile(csvGCPFile, csvPath, processedPath);
   }
