+++ src/test/java/uk/gov/ons/fwmt/csvservice/service/CCSConverterServiceImplTest.java
@@ -11,8 +11,8 @@ import org.springframework.core.io.Resource;
 import org.springframework.core.io.WritableResource;
 import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.ons.census.fwmt.common.error.GatewayException;
-import uk.gov.ons.census.fwmt.csvservice.service.impl.CCSConverterServiceImpl;
-import uk.gov.ons.census.fwmt.csvservice.service.impl.CSVAdapterServiceImpl;
+import uk.gov.ons.census.fwmt.csvservice.service.ce.CCSConverterServiceImpl;
+import uk.gov.ons.census.fwmt.csvservice.service.ce.GatewayActionAdapter;
 import uk.gov.ons.census.fwmt.events.component.GatewayEventManager;
 
 import java.io.OutputStream;
@@ -29,7 +29,7 @@ public class CCSConverterServiceImplTest {
   private CCSConverterServiceImpl ccsConverterService;
 
   @Mock
-  private CSVAdapterServiceImpl csvAdapterService;
+  private GatewayActionAdapter gatewayActionAdapter;
 
   @Mock
   private GatewayEventManager gatewayEventManager;
